+++ sql_hana.go
@@ -90,14 +90,14 @@ func BuildConnection(
 			if strings.HasPrefix(passwd, "/encrypted/") {
 				if val, auth_key, err := BuildPasswd(logger, passwd, symbol_table); err == nil {
 					params["password"] = val
-					params["auth_key"] = auth_key
+					params["__auth_key"] = auth_key
 				} else {
 					return "", fmt.Errorf("unable to decrypt password")
 				}
-				params["need_auth_key"] = "true"
+				params["__need_auth_key"] = "true"
 			} else {
 				params["password"] = val
-				params["need_auth_key"] = "false"
+				params["__need_auth_key"] = "false"
 			}
 
 			// add params to target symbol table